refactor(main): split filesystem git remote handlers
(cherry picked from commit 2146cff06aad7998ecfd5b65b853576e6757c866)
