+++ src/main/ipc/filesystem/filesystem-git-remote-handlers.ts
@@ -1,322 +1,10 @@
-import { ipcMain } from 'electron'
-import type {
-  GitBranchCompareResult,
-  GitCommitCompareResult
-} from '../../../shared/git-diff-compare-types'
-import type { GitForkSyncExpectedUpstream, GitForkSyncResult } from '../../../shared/git-fork-sync'
-import type { GitUpstreamStatus } from '../../../shared/git-status-types'
-import type { GitPushTarget } from '../../../shared/worktree/types'
-import type { GitAdmissionTier } from '../../git/command-runner/git-exec-options'
-import { getBranchCompare, getCommitCompare } from '../../git/status'
-import { gitFetch, gitPush, gitPull, gitFastForward, gitPullRebaseFromBase } from '../../git/remote'
-import { gitSyncForkDefaultBranch } from '../../git/fork-sync'
-import { getUpstreamStatus } from '../../git/upstream'
-import {
-  getSshGitProvider,
-  SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE
-} from '../../providers/ssh-git-dispatch'
-import { resolveRegisteredWorktreePath } from '../registered-worktree-roots-cache'
-import { getLocalGitOptionsForRegisteredWorktree } from '../local-worktree-runtime-options'
-import { validateGitPushTarget } from '../../git/push-target-validation'
-import { assertGitPushTargetShape } from '../../../shared/git-push-target-validation'
-import { validateGitForkSyncExpectedUpstream } from '../../../shared/git-fork-sync'
-import { validateFullGitObjectId } from './filesystem-worktree-helpers'
 import type { FilesystemHandlerContext } from './filesystem-handler-context'
+import { registerGitRemoteBranchMutationHandlers } from './git-remote/branch-mutation-handlers'
+import { registerGitRemoteCompareHandlers } from './git-remote/compare-handlers'
+import { registerGitRemoteSyncHandlers } from './git-remote/sync-handlers'
 
 export function registerFilesystemGitRemoteHandlers(context: FilesystemHandlerContext): void {
-  const { store } = context
-  ipcMain.handle(
-    'git:branchCompare',
-    async (
-      _event,
-      args: {
-        worktreePath: string
-        baseRef: string
-        connectionId?: string
-        admissionTier?: GitAdmissionTier
-      }
-    ): Promise<GitBranchCompareResult> => {
-      if (args.connectionId) {
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return args.admissionTier
-          ? provider.getBranchCompare(args.worktreePath, args.baseRef, {
-              admissionTier: args.admissionTier
-            })
-          : provider.getBranchCompare(args.worktreePath, args.baseRef)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      return getBranchCompare(worktreePath, args.baseRef, {
-        ...gitOptions,
-        ...(args.admissionTier ? { admissionTier: args.admissionTier } : {})
-      })
-    }
-  )
-
-  ipcMain.handle(
-    'git:commitCompare',
-    async (
-      _event,
-      args: { worktreePath: string; commitId: string; connectionId?: string }
-    ): Promise<GitCommitCompareResult> => {
-      const commitId = validateFullGitObjectId(args.commitId, 'commitId')
-      if (args.connectionId) {
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.getCommitCompare(args.worktreePath, commitId)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      return getCommitCompare(worktreePath, commitId, gitOptions)
-    }
-  )
-
-  ipcMain.handle(
-    'git:upstreamStatus',
-    async (
-      _event,
-      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
-    ): Promise<GitUpstreamStatus> => {
-      if (args.connectionId) {
-        if (args.pushTarget) {
-          assertGitPushTargetShape(args.pushTarget)
-        }
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.getUpstreamStatus(args.worktreePath, args.pushTarget)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      return getUpstreamStatus(worktreePath, args.pushTarget, gitOptions)
-    }
-  )
-
-  ipcMain.handle(
-    'git:fetch',
-    async (
-      _event,
-      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
-    ): Promise<void> => {
-      if (args.connectionId) {
-        if (args.pushTarget) {
-          assertGitPushTargetShape(args.pushTarget)
-        }
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.fetchRemote(args.worktreePath, args.pushTarget)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      if (args.pushTarget) {
-        await validateGitPushTarget(worktreePath, args.pushTarget, {
-          ...gitOptions,
-          admissionTier: 'interactive'
-        })
-      }
-      await gitFetch(worktreePath, args.pushTarget, {
-        ...gitOptions,
-        admissionTier: 'interactive'
-      })
-    }
-  )
-
-  ipcMain.handle(
-    'git:syncFork',
-    async (
-      _event,
-      args: {
-        worktreePath: string
-        connectionId?: string
-        expectedUpstream: GitForkSyncExpectedUpstream
-      }
-    ): Promise<GitForkSyncResult> => {
-      const expectedUpstream = validateGitForkSyncExpectedUpstream(args.expectedUpstream, {
-        required: true
-      })
-      if (args.connectionId) {
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.syncForkDefaultBranch(args.worktreePath, expectedUpstream)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      return gitSyncForkDefaultBranch(worktreePath, expectedUpstream, {
-        ...gitOptions,
-        admissionTier: 'interactive'
-      })
-    }
-  )
-
-  ipcMain.handle(
-    'git:push',
-    async (
-      _event,
-      args: {
-        worktreePath: string
-        publish?: boolean
-        forceWithLease?: boolean
-        connectionId?: string
-        pushTarget?: GitPushTarget
-      }
-    ): Promise<void> => {
-      // Why: coerce to strict boolean so a malformed payload (e.g. string 'false') can't enable --set-upstream; mirror in src/relay/git-handler.ts.
-      const publish = args.publish === true
-      if (args.connectionId) {
-        if (args.pushTarget) {
-          assertGitPushTargetShape(args.pushTarget)
-        }
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.pushBranch(args.worktreePath, publish, args.pushTarget, {
-          forceWithLease: args.forceWithLease === true
-        })
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      if (args.pushTarget) {
-        await validateGitPushTarget(worktreePath, args.pushTarget, {
-          ...gitOptions,
-          admissionTier: 'interactive'
-        })
-      }
-      await gitPush(worktreePath, publish, args.pushTarget, {
-        forceWithLease: args.forceWithLease === true,
-        ...gitOptions,
-        admissionTier: 'interactive'
-      })
-    }
-  )
-
-  ipcMain.handle(
-    'git:pull',
-    async (
-      _event,
-      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
-    ): Promise<void> => {
-      if (args.connectionId) {
-        if (args.pushTarget) {
-          assertGitPushTargetShape(args.pushTarget)
-        }
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.pullBranch(args.worktreePath, args.pushTarget)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      if (args.pushTarget) {
-        await validateGitPushTarget(worktreePath, args.pushTarget, {
-          ...gitOptions,
-          admissionTier: 'interactive'
-        })
-      }
-      await gitPull(worktreePath, args.pushTarget, {
-        ...gitOptions,
-        admissionTier: 'interactive'
-      })
-    }
-  )
-
-  ipcMain.handle(
-    'git:fastForward',
-    async (
-      _event,
-      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
-    ): Promise<void> => {
-      if (args.connectionId) {
-        if (args.pushTarget) {
-          assertGitPushTargetShape(args.pushTarget)
-        }
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.fastForwardBranch(args.worktreePath, args.pushTarget)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      if (args.pushTarget) {
-        await validateGitPushTarget(worktreePath, args.pushTarget, {
-          ...gitOptions,
-          admissionTier: 'interactive'
-        })
-      }
-      await gitFastForward(worktreePath, args.pushTarget, {
-        ...gitOptions,
-        admissionTier: 'interactive'
-      })
-    }
-  )
-
-  ipcMain.handle(
-    'git:rebaseFromBase',
-    async (
-      _event,
-      args: { worktreePath: string; baseRef: string; connectionId?: string }
-    ): Promise<void> => {
-      if (args.connectionId) {
-        const provider = getSshGitProvider(args.connectionId)
-        if (!provider) {
-          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
-        }
-        return provider.rebaseFromBase(args.worktreePath, args.baseRef)
-      }
-      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
-      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
-        store,
-        args.worktreePath,
-        worktreePath
-      )
-      await gitPullRebaseFromBase(worktreePath, args.baseRef, {
-        ...gitOptions,
-        admissionTier: 'interactive'
-      })
-    }
-  )
+  registerGitRemoteCompareHandlers(context)
+  registerGitRemoteSyncHandlers(context)
+  registerGitRemoteBranchMutationHandlers(context)
 }
