+++ src/main/ipc/filesystem/git-remote/branch-mutation-handlers.ts
@@ -1,0 +1,158 @@
+import { ipcMain } from 'electron'
+import type { GitPushTarget } from '../../../../shared/worktree/types'
+import { gitFastForward, gitPull, gitPullRebaseFromBase, gitPush } from '../../../git/remote'
+import { validateGitPushTarget } from '../../../git/push-target-validation'
+import {
+  getSshGitProvider,
+  SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE
+} from '../../../providers/ssh-git-dispatch'
+import { resolveRegisteredWorktreePath } from '../../registered-worktree-roots-cache'
+import { getLocalGitOptionsForRegisteredWorktree } from '../../local-worktree-runtime-options'
+import { assertGitPushTargetShape } from '../../../../shared/git-push-target-validation'
+import type { FilesystemHandlerContext } from '../filesystem-handler-context'
+
+export function registerGitRemoteBranchMutationHandlers(context: FilesystemHandlerContext): void {
+  const { store } = context
+
+  ipcMain.handle(
+    'git:push',
+    async (
+      _event,
+      args: {
+        worktreePath: string
+        publish?: boolean
+        forceWithLease?: boolean
+        connectionId?: string
+        pushTarget?: GitPushTarget
+      }
+    ): Promise<void> => {
+      // Why: coerce to strict boolean so a malformed payload (e.g. string 'false') can't enable --set-upstream; mirror in src/relay/git-handler.ts.
+      const publish = args.publish === true
+      if (args.connectionId) {
+        if (args.pushTarget) {
+          assertGitPushTargetShape(args.pushTarget)
+        }
+        const provider = getSshGitProvider(args.connectionId)
+        if (!provider) {
+          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
+        }
+        return provider.pushBranch(args.worktreePath, publish, args.pushTarget, {
+          forceWithLease: args.forceWithLease === true
+        })
+      }
+      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
+      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
+        store,
+        args.worktreePath,
+        worktreePath
+      )
+      if (args.pushTarget) {
+        await validateGitPushTarget(worktreePath, args.pushTarget, {
+          ...gitOptions,
+          admissionTier: 'interactive'
+        })
+      }
+      await gitPush(worktreePath, publish, args.pushTarget, {
+        forceWithLease: args.forceWithLease === true,
+        ...gitOptions,
+        admissionTier: 'interactive'
+      })
+    }
+  )
+
+  ipcMain.handle(
+    'git:pull',
+    async (
+      _event,
+      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
+    ): Promise<void> => {
+      if (args.connectionId) {
+        if (args.pushTarget) {
+          assertGitPushTargetShape(args.pushTarget)
+        }
+        const provider = getSshGitProvider(args.connectionId)
+        if (!provider) {
+          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
+        }
+        return provider.pullBranch(args.worktreePath, args.pushTarget)
+      }
+      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
+      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
+        store,
+        args.worktreePath,
+        worktreePath
+      )
+      if (args.pushTarget) {
+        await validateGitPushTarget(worktreePath, args.pushTarget, {
+          ...gitOptions,
+          admissionTier: 'interactive'
+        })
+      }
+      await gitPull(worktreePath, args.pushTarget, {
+        ...gitOptions,
+        admissionTier: 'interactive'
+      })
+    }
+  )
+
+  ipcMain.handle(
+    'git:fastForward',
+    async (
+      _event,
+      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
+    ): Promise<void> => {
+      if (args.connectionId) {
+        if (args.pushTarget) {
+          assertGitPushTargetShape(args.pushTarget)
+        }
+        const provider = getSshGitProvider(args.connectionId)
+        if (!provider) {
+          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
+        }
+        return provider.fastForwardBranch(args.worktreePath, args.pushTarget)
+      }
+      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
+      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
+        store,
+        args.worktreePath,
+        worktreePath
+      )
+      if (args.pushTarget) {
+        await validateGitPushTarget(worktreePath, args.pushTarget, {
+          ...gitOptions,
+          admissionTier: 'interactive'
+        })
+      }
+      await gitFastForward(worktreePath, args.pushTarget, {
+        ...gitOptions,
+        admissionTier: 'interactive'
+      })
+    }
+  )
+
+  ipcMain.handle(
+    'git:rebaseFromBase',
+    async (
+      _event,
+      args: { worktreePath: string; baseRef: string; connectionId?: string }
+    ): Promise<void> => {
+      if (args.connectionId) {
+        const provider = getSshGitProvider(args.connectionId)
+        if (!provider) {
+          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
+        }
+        return provider.rebaseFromBase(args.worktreePath, args.baseRef)
+      }
+      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
+      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
+        store,
+        args.worktreePath,
+        worktreePath
+      )
+      await gitPullRebaseFromBase(worktreePath, args.baseRef, {
+        ...gitOptions,
+        admissionTier: 'interactive'
+      })
+    }
+  )
+}
