+++ src/main/ipc/filesystem/git-remote/sync-handlers.ts
@@ -1,0 +1,118 @@
+import { ipcMain } from 'electron'
+import type {
+  GitForkSyncExpectedUpstream,
+  GitForkSyncResult
+} from '../../../../shared/git-fork-sync'
+import type { GitPushTarget } from '../../../../shared/worktree/types'
+import type { GitUpstreamStatus } from '../../../../shared/git-status-types'
+import { gitFetch } from '../../../git/remote'
+import { gitSyncForkDefaultBranch } from '../../../git/fork-sync'
+import { getUpstreamStatus } from '../../../git/upstream'
+import { validateGitPushTarget } from '../../../git/push-target-validation'
+import {
+  getSshGitProvider,
+  SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE
+} from '../../../providers/ssh-git-dispatch'
+import { resolveRegisteredWorktreePath } from '../../registered-worktree-roots-cache'
+import { getLocalGitOptionsForRegisteredWorktree } from '../../local-worktree-runtime-options'
+import { assertGitPushTargetShape } from '../../../../shared/git-push-target-validation'
+import { validateGitForkSyncExpectedUpstream } from '../../../../shared/git-fork-sync'
+import type { FilesystemHandlerContext } from '../filesystem-handler-context'
+
+export function registerGitRemoteSyncHandlers(context: FilesystemHandlerContext): void {
+  const { store } = context
+
+  ipcMain.handle(
+    'git:upstreamStatus',
+    async (
+      _event,
+      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
+    ): Promise<GitUpstreamStatus> => {
+      if (args.connectionId) {
+        if (args.pushTarget) {
+          assertGitPushTargetShape(args.pushTarget)
+        }
+        const provider = getSshGitProvider(args.connectionId)
+        if (!provider) {
+          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
+        }
+        return provider.getUpstreamStatus(args.worktreePath, args.pushTarget)
+      }
+      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
+      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
+        store,
+        args.worktreePath,
+        worktreePath
+      )
+      return getUpstreamStatus(worktreePath, args.pushTarget, gitOptions)
+    }
+  )
+
+  ipcMain.handle(
+    'git:fetch',
+    async (
+      _event,
+      args: { worktreePath: string; connectionId?: string; pushTarget?: GitPushTarget }
+    ): Promise<void> => {
+      if (args.connectionId) {
+        if (args.pushTarget) {
+          assertGitPushTargetShape(args.pushTarget)
+        }
+        const provider = getSshGitProvider(args.connectionId)
+        if (!provider) {
+          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
+        }
+        return provider.fetchRemote(args.worktreePath, args.pushTarget)
+      }
+      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
+      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
+        store,
+        args.worktreePath,
+        worktreePath
+      )
+      if (args.pushTarget) {
+        await validateGitPushTarget(worktreePath, args.pushTarget, {
+          ...gitOptions,
+          admissionTier: 'interactive'
+        })
+      }
+      await gitFetch(worktreePath, args.pushTarget, {
+        ...gitOptions,
+        admissionTier: 'interactive'
+      })
+    }
+  )
+
+  ipcMain.handle(
+    'git:syncFork',
+    async (
+      _event,
+      args: {
+        worktreePath: string
+        connectionId?: string
+        expectedUpstream: GitForkSyncExpectedUpstream
+      }
+    ): Promise<GitForkSyncResult> => {
+      const expectedUpstream = validateGitForkSyncExpectedUpstream(args.expectedUpstream, {
+        required: true
+      })
+      if (args.connectionId) {
+        const provider = getSshGitProvider(args.connectionId)
+        if (!provider) {
+          throw new Error(SSH_GIT_PROVIDER_UNAVAILABLE_MESSAGE)
+        }
+        return provider.syncForkDefaultBranch(args.worktreePath, expectedUpstream)
+      }
+      const worktreePath = await resolveRegisteredWorktreePath(args.worktreePath, store)
+      const gitOptions = getLocalGitOptionsForRegisteredWorktree(
+        store,
+        args.worktreePath,
+        worktreePath
+      )
+      return gitSyncForkDefaultBranch(worktreePath, expectedUpstream, {
+        ...gitOptions,
+        admissionTier: 'interactive'
+      })
+    }
+  )
+}
